+++ R/plot_plotly.R
@@ -4,7 +4,7 @@
 #' @param linewidth the width in pixel for the range lines
 #' @param title the title for the plot
 #' @param show_labels boolean, show labels on events or not
-#' @param background_lines number of grey background lines to draw
+#' @param background_lines number of grey background lines to draw (can be NULL)
 #'
 #' @return a list containing the plots for the groups in data
 #' @keywords internal
@@ -23,29 +23,11 @@ plot_plotly <- function(data, linewidth, title, show_labels, background_lines) {
 
   # 1. Prepare basic plot
   p <- plotly::plot_ly(type = "scatter", mode = "lines")
-  maxY <- max(data$y) + 1
 
-  # 1. add vertical lines
-  for (day in seq(min(c(data$start, data$end)), max(c(data$start, data$end)), length.out = background_lines + 1)) {
-    p <- plotly::add_trace(p,
-                           x = as.POSIXct(day, origin = "1970-01-01"), y = c(0, maxY), mode = "lines",
-                           line = list(color = plotly::toRGB("grey90")), showlegend = F, hoverinfo = "none"
-    )
-  }
-
-  # 2. Divide subplots with horizontal lines
-  divide_at_y <- setdiff(seq_len(max(data$y)), data$y)
-  for (i in  divide_at_y) {
-    p <- plotly::add_trace(p,
-                           x = c(min(data$start), max(data$end)), y =i, mode = "lines",
-                           line = list(color = plotly::toRGB("grey65")), showlegend = F, hoverinfo = "none"
-    )
-  }
-
-  # 1. plot ranges
+  # 2. plot ranges
   range_dat <- data[data$start != data$end, ]
 
-  linewidth <- ifelse(is.null(linewidth), max(-3 * max(data$subplot) + max(data$y) + 20, 15), linewidth)
+  lw <- ifelse(is.null(linewidth), 300 / max(data$y), linewidth)
 
   if(nrow(range_dat) > 0){
     # draw ranges piecewise
@@ -55,7 +37,7 @@ plot_plotly <- function(data, linewidth, title, show_labels, background_lines) {
       p <- plotly::add_trace(p,
                              x = c(toAdd$start, toAdd$end), # von, bis
                              y = toAdd$y,
-                             line = list(color = toAdd$col, width = linewidth),
+                             line = list(color = toAdd$col, width = lw),
                              showlegend = F,
                              hoverinfo = "text",
                              text = toAdd$tooltip
@@ -75,7 +57,7 @@ plot_plotly <- function(data, linewidth, title, show_labels, background_lines) {
     }
   }
 
-  # 2. plot events
+  # 3. plot events
   event_dat <- data[data$start == data$end, ]
   if(nrow(event_dat) > 0){
     # alternate y positions for event labels
@@ -85,7 +67,7 @@ plot_plotly <- function(data, linewidth, title, show_labels, background_lines) {
     p <- plotly::add_markers(p,
                              x = event_dat$start, y = event_dat$y,
                              marker = list(
-                               color = event_dat$col, size = 15, symbol = "circle",
+                               color = event_dat$col, size = 0.7 * lw, symbol = "circle",
                                line = list(color = "black", width = 1)
                              ),
                              showlegend = F, hoverinfo = "text", text = event_dat$tooltip
@@ -104,15 +86,32 @@ plot_plotly <- function(data, linewidth, title, show_labels, background_lines) {
 
   y_ticks <- sapply(split(data, data$subplot), function(subplot) mean(subplot$y))
 
+  # 4. Divide subplots with horizontal lines
+  hline <- function(y = 0) list(type = "line", x0 = 0, x1 = 1, xref = "paper", y0 = y, y1 = y, line = list(color = "grey65", width = 0.5))
+  vline <- function(x = 0) list(type = "line", y0 = 0, y1 = 1, yref = "paper", x0 = x, x1 = x, line = list(color = "grey90", width = 0.1))
+  horizontal_lines <- lapply(setdiff(seq_len(max(data$y)), data$y), hline)
+
+  # 4. Add vertical lines
+  if(!is.null(background_lines)){
+    day_breaks <- as.POSIXct(seq(min(c(data$start, data$end)), max(c(data$start, data$end)),
+                                     length.out = round(background_lines) + 2), origin = "1970-01-01")
+    vertical_lines <- lapply(day_breaks, vline)
+  }else{
+    vertical_lines <- list()
+  }
+
   p <- plotly::layout(p,
                       hovermode = "closest",
+                      plot_bgcolor = "#FCFCFC",
                       title = title,
-                      margin = list(l = max(nchar(as.character(data$group))) * 8),
+                      shapes = append(vertical_lines, horizontal_lines),
                       # Axis options:
-                      # 1. Remove gridlines
-                      # 2. Customize y-axis tick labels and show group names instead of numbers
-                      xaxis = list(showgrid = F, title = ""),
+                      xaxis = list(linewidth = 1,  mirror = TRUE,
+                                   showgrid = is.null(background_lines),
+                                   gridcolor = "grey90", title = ""),
                       yaxis = list(
+                        linewidth = 1, mirror = TRUE,
+                        range = c(0, max(data$y) + 1),
                         showgrid = F, title = "",
                         tickmode = "array",
                         tickvals = y_ticks,